*** empty log message ***
git-svn-id: https://qucs.svn.sourceforge.net/svnroot/qucs/trunk@1076 b5b04e8c-4942-46c9-ab4f-83783d557d1c

+++ qucs/qucs/diagrams/polardiagram.h
@@ -24,14 +24,14 @@
 class PolarDiagram : public Diagram  {
 public: 
   PolarDiagram(int _cx=0, int _cy=0);
-  ~PolarDiagram();
+ ~PolarDiagram();
 
 
   Diagram* newOne();
   static Element* info(QString&, char* &, bool getNewOne=false);
   int  calcDiagram();
   void calcLimits();
-  void calcCoordinate(double* &, double* &, double* &, int*, int*, Axis*);
+  void calcCoordinate(double* &, double* &, double* &, float*, float*, Axis*);
 };
 
 #endif
